move to seperate class
